Don't trigger a close due to a release if we did not get a press
If the drawer does not see the press point then it was not open at the
time it was pressed. This ensures that when a touch is done on a button
to trigger the opening of a drawer it does not think it should be
closed due to the default close policy.

Pick-to: 5.15
Change-Id: I0d1b11341e659dc14554d0d72aac70199bbb99fa
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>