Fix QProperty property interaction with aliases
With this change, an alias of a bindable property is also bindable, and
shares its bindable interface with the target.

Moreover, the logic in QQmlTypeCompiler is adjusted so that a change
handler of an alias uses the bindable interface if possible, instead of
connecting to the alias' change signal. That would never be emitted if
the target is a QProperty without a notify signal.

Alias properties still have a change signal, but those never get
emitted.

Change-Id: I857dfdbe51048a2b604ad632982e7f4adac6b907
Reviewed-by: Lars Knoll <lars.knoll@qt.io>