QQmlProperty: Guard QQmlContextData with QQmlGuardedContextData
Some animation classes store the property to animate in a
QQmlProperty.
In some cases (e.g. ListView delegates), the item that is animated
(and therefore the context) might get destroyed while the animation
is running. And since the QQmlContextData in QQmlProperty was not
guarded, it could sometimes cause a access violation crash as the
task demonstrates.

Fixes: QTBUG-86567
Change-Id: I2c8c4f0dca8029887702adeae542eadc47451f60
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>