Avoid crash when module version cannot be resolved
When QQmlMetaType::latestModuleVersion() returned an invalid
version, we would fail importStaticPlugin() without adding
any error.

Since we unconditionally called takeFirst() on
the errors list, we would later crash. The precondition
here is that any failure adds an error, so we now assert to
make this explicit. Also, add a simple module not found
error when we can't resolve it.

Change-Id: I7a13775b2ef56d95b7738ac53e58501b8c68895f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>