Adjust testbench to properly support multiple styles
We need to split the uses of the Imagine style API out into
file-selected files. We set Imagine.path so that the default
assets are used for areas of the UI that are important to the
functioning of the example, as using custom assets here could
make it impossible to use the API if the assets were not
created correctly.

Change-Id: I2b6b683ffe65b7d573abf0ae793509a95417aead
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>