NativeStyle: add focus frame to the macOS style
This patch will add a focus frame to the macOS style.
We use one focus frame item that we move around according
to changes to the focus object.

It's normally the background delegate's responsibility to
draw a focus frame, so we take care to only draw the
native focus frame for controls with default delegates.
(to be consistent with the other styles)

Change-Id: Iaa7202675c1aad2fc19682563ac1afb6e686c105
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>