QQItemPriv::localizedTouchEvent: don't skip parents of passive grabbers
Some subtests of tst_FlickableInterop::touchAndDragHandlerOnFlickable()
check whether Flickable can continue to filter child mouse events, even
when the only thing going on is that we are delivering a TouchUpdate
QTouchEvent to a TapHandler inside the Flickable's content item (such as
inside a ListView delegate) due to its having a passive grab.  So during
filtering only, localizedTouchEvent() needs to localize the event for
any item that has passive grabbers somewhere in its children.

Change-Id: Icff0438ddba97bd916feedc9dcd6a71ae47671e5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>