PathView: stop using QQWindowPriv::cloneMouseEvent; refactor filtering
There's no need to heap-allocate a short-term mouse event; and
localizing one is easy enough via QMutableEventPoint::setPosition.
The sendMouseEvent() function (and its name) didn't make as much sense
as handling the filtering completely in the childMouseEventFilter()
function.  Checking and changing the grabber is also now possible via
accessors in the event itself, so now we have one fewer use of
QQuickItem::grabMouse() and QQuickWindow::mouseGrabberItem().

Change-Id: I84f9f26e349a8d1aabacdbdbb264949d1103e91d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>