Prevent touch->mouse synth from altering the localized touchpoint
QEventPoint is explicitly shared, so a QMouseEvent synthesized from
a QTouchEvent shared the same QEventPoint instance.
When QQuickFlickable::filterMouseEvent() calls cloneMouseEvent(),
it was re-localizing the point for the flickable, after it had
already been localized for delivery to the original receiver item.
This caused a lot of failures in Controls, e.g. for any button inside
a Flickable, QQuickAbstractButtonPrivate::handlePress() would be given
the wrong position.

After filtering, we need to be able to resume delivery to the
original receiver item without re-localizing the point.

During filtering, the filtering parent should receive the same version
of the touch event that contains only the points that would be sent
to the receiver item, not the potentially more-complete original event.

Fixes: QTBUG-87157
Change-Id: I7eec6f5ecfe9f042199f0944897c04fbffb2172e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>