QQmlPropertyValidator: only use string converter for strings
A recent change in fromIsoTimeString exposed the fact that we are using
QML's string converters to check whether a binding expression is valid
for a given property. However, while we begrudgingly accept that string
representations shall be converted to value types via the string
converters, the same shouldn't hold true for things that aren't actually
strings - for instance, numbers.

Fixes: QTBUG-87299
Change-Id: Iefd390efae7c193dc32d37e63943b39e09c9295a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>