Native style, macOS: don't draw a background behind the GroupBox label
On macOS, the controls are semi-transparent. This means
that if you use a different color for e.g the window
background, all the controls will get a different
color too.

For this reason, we should avoid drawing a background
color for the GroupBox label. And to achieve that, we
need to override the default background set from
DefaultGroupBox, to use an Item instead of a Rectangle.

Change-Id: I60c496c19744763adb2ec683670c4a9c64cc32fa
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>