Native style, macOS: always draw a TextField using NSTextField
There is no apparent reason as to why we should let QCommonStyle
fill the background before it calls PE_FrameLineEdit in QMacStyle.
The NSTextField cell already draws the correct background, which
is supposede to be semi-transparent in dark-mode.

Therefore, just let PE_PanelLineEdit and PE_FrameLineEdit be
drawn the same way.

Change-Id: Iacd4f1b1802e1b0c2b96a92cf271b681c10b8cff
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>