macOS/ScrollView: don't create contentItem too early
When creating the binding "contentItem.clip:" we
accidentally also created the contentItem before
the component was completed. Since the contentItem
of a ScrollView is a Flickable, it meant that we
ended up always creating an internal flickable
from QQuickScrollViewPrivate::ensureFlickable,
instead of using the Flickable child if there
was one. So this binding cannot be used.

Change-Id: I2e7ca63f64693adb1b524966fec82aef2b3fb658
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>