QQuickShaderEffect: fix crash when hiding parent
It's possible for itemChange to be called during destruction when
deleting the QQuickShaderEffectImpl. We nullify m_impl before deleting
it via another pointer to it, so we must check that it's not null
before trying to use it.

Fixes: QTBUG-86402
Change-Id: If4955445f7cc0d1f376bc9b86b95e1cca4d88ede
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 266cd7638d887b31d56964a0f13fe208821703b1)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>