Deal with QEvent::setAccepted() calling QEventPoint::setAccepted()
Traditionally, a QQuickItem subclass could override touchEvent();
if the event is still accepted when it returns, that meant the item
handled all the points, so completely that the item should get an
implicit grab of those points, and they should not be delivered further
to any other items underneath.  But it's often not so simple in
multi-touch UIs, which is why each QEventPoint has its own accept flag.

QPointerEvent::setAccepted() now iterates the points and calls
QEventPoint::setAccepted() on each; so QQuickWindow doesn't need to do
anything to make that happen anymore.

Change-Id: Idbe7abf278411ee2fea598c1a1939f4bdb214aa0
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>