nativestyle, ScrollView: don't resize scroll bars according to padding
Depending on if you see the scroll bars as a part of the
contents, the background, or neither, they should respect
either padding, insets, or none of them. The styles that
use transient scrollbars see them partly as a part of the
contents (judging from their implementation), which means
that their width and height will respect padding, but
their position will not (they will always be glued to
the right or the bottom side of the control).

Letting width and height be affected by padding doesn't
work so well with non-transient scroll bars. In that
case the scroll bars should normally be "connected"
at the bottom-right edge. So change the code so that
the scrollbars ignore padding, and instead use the full
width and height of the control.

Change-Id: Ifa5a7708a51b3773a63ebdd50781eb3845be4744
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>