Register native styles under 2.0 in addition to 6.0
Since we changed the logic that decides the default style, users with
QML that imports 2.x versions of QtQuick.Controls will get errors if
the default style that is chosen is a native style:

qrc:/main.qml:53:1: module "QtQuick.Controls.macOS" version 2.15 is not installed
qrc:/main.qml: Failed to load dependencies for module "QtQuick.Controls" version 2.15

We use PAST_MAJOR_VERSIONS to register the module for 2.x, and not
just 6.0, and register the types of those styles under version 2.0
in addition to 6.0. This should not cause any problems for existing
QML code.

Fixes: QTBUG-87658
Change-Id: I976078c0aea3fbebe4e04f5ba4d8056c0ca7a0e6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>