Make tst_QQuickMultiPointTouchArea::inFlickable() more debuggable
It was all-white despite having two items that react to events, which
makes it hard to see what's failing when it fails.

Use loops rather than tedious repetitive hard-coded moves to improve
readability.

"ensure that mouse events do not fall through to the Flickable"
must include having the MPTA "eat" the mouse press.

Change-Id: Ibca14709afbd561f03a62d74d10e6cae94ab302e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>