Modernize event handling in PinchArea
Worrying about the window's touchmouse is really not a relevant concern
for a touch-only item; and grabs are done by grabbing the specific
points that the pincharea chooses to handle, not by grabbing the
touchmouse.  We simply didn't have suitable API in Qt 5.

Also some drive-by fixes: better packing and initialization in
QQuickPinchAreaPrivate; and the stealMouse variable has only
been set, not checked, for a long time, so we can remove it.
Remove unimplemented handlePress() and handleRelease().

Change-Id: Ia76dc9b9974f663b13c4bb9dac32efe8ed7815c5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>